Update README.md
In this scenario, the issue that I was tasked with fixing was helping a company called Chada Tech with their clock. Chada tech wanted me as a computer scientist to create an international clock. This would allow their clients to be able to view a 12 and a 24-hour clock on their website rather than just the 12-hour clock. 

I would say that what I did particularly well was my pseudocode which I do sometimes struggle with, and I feel like I also did very well with my actual code itself, well to an extent. Now, I could have enhanced my code because when reading it, it didn't seem like there was an issue. However, when it came to us using the code, it wouldn't display the clock properly on my screen. When I turned it in, it didn't show the clock at all on my professor screen, but my code itself didn’t appear to have any issues. 

The part of the code I found the most difficult to write was where we had to put each clock in their own separate box.  I feel like I couldn't get either one of the clocks to align properly, so when I viewed it, it was either flat, shifted too far to the left or shifted too far to the right, but never perfectly centered 

I think everything in this project would be particularly transferable to other projects. We had to do a little bit of math, a bit of aesthetics of our clock, but a lot of coding. In the future, I feel like that's going to help me a lot, especially since C++ is a program that is used within intricate development